refactor(name): rename gpg to pgp
