+++ core/src/operation/init.rs
@@ -4,15 +4,15 @@ use std::path::PathBuf;
 
 use secrecy::ExposeSecret;
 
-use crate::gpg::utils::{check_recipient_type, user_email_to_fingerprint, RecipientType};
-use crate::gpg::GPGClient;
+use crate::pgp::utils::{check_recipient_type, user_email_to_fingerprint, RecipientType};
+use crate::pgp::PGPClient;
 use crate::util::fs_utils::{backup_encrypted_file, path_to_str, process_files_recursively};
 use crate::{IOErr, IOErrType};
 
 const FPR_FILENAME: &str = ".gpg-id";
 
 pub fn init(
-    client: &GPGClient,
+    client: &PGPClient,
     root_path: &PathBuf,
     target_path: &str,
     recipient: &str,
