+++ core/src/operation/show.rs
@@ -1,1 +1,9 @@
+use std::path::Path;
 
+use crate::pgp::PGPClient;
+
+
+pub fn grep_string(gpg_client:&PGPClient, root:&Path,to_search:&str)->Vec<String>{
+  todo!()
+
+}
